Use a project specific restore directory
It's probably safer to use a project specific restore directory by default
rather than a single top-level restore directory which is relative to this
file.

This way if multiple projects are set up to use corebuild.props they
will generate an independent .nuget directory in the same dir as
the project itself.