Switch to using the logical view of the project
This significantly improves the correctness of the help, since
it does not rely on evaluated conditions that may be false when
discovering properties and targets.

Also improved the comment location by considering all elements
that declare the same property/target, and concatenating their
documentation automatically, instead of getting just the first.