Explicitly override potential binlog specified in autoresponse
The previous fix caused an infinite loop call since by not loading the autoresponse file, the 
`ColoredConsole` property would never be true and therefore we'd keep calling `_ConfigureColoredLogger` 
infinitely. 

The proper fix for the scenario where there is already an `msbuild.rsp` file that has the `/bl` switch is to actually 
skip the automatic loading, then loading it explicitly with `@msbuild.rsp` (we know it will exist since the previous 
task would have generated it), then override the `/bl` switch by providing our own, and finally deleting the 
transient binlog we generated, since it's not actually needed.

Also, the final message after configuration could have gone missing in the output if there was a `/v:m` verbosity 
setting specified too.